[Async Salable Status] Moved product salability calculation from observer to plugin method to be executed on demand.
