+++ InventoryBundleProduct/etc/di.xml
@@ -26,7 +26,4 @@
             </argument>
         </arguments>
     </type>
-    <type name="Magento\CatalogInventory\Helper\Stock">
-        <plugin name="adapt_assign_stock_status_to_bundle_product" type="Magento\InventoryBundleProduct\Plugin\CatalogInventory\Helper\Stock\AdaptAssignStatusToProductPlugin"/>
-    </type>
 </config>
