+++ InventoryCatalog/Model/IsProductSalable.php
@@ -1,0 +1,67 @@
+<?php
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+
+declare(strict_types=1);
+
+namespace Magento\InventoryCatalog\Model;
+
+use Magento\Catalog\Model\Product;
+use Magento\InventoryCatalogApi\Model\GetProductIdsBySkusInterface;
+use Magento\InventorySalesApi\Api\AreProductsSalableInterface;
+
+class IsProductSalable
+{
+    /**
+     * @var GetStockIdForCurrentWebsite
+     */
+    private $getStockIdForCurrentWebsite;
+
+    /**
+     * @var AreProductsSalableInterface
+     */
+    private $areProductsSalable;
+
+    /**
+     * @var array
+     */
+    private $productStatusCache;
+
+    /**
+     * @param GetStockIdForCurrentWebsite $getStockIdForCurrentWebsite
+     * @param AreProductsSalableInterface $areProductsSalable
+     * @param GetProductIdsBySkusInterface $getProductIdsBySkus
+     */
+    public function __construct(
+        GetStockIdForCurrentWebsite $getStockIdForCurrentWebsite,
+        AreProductsSalableInterface $areProductsSalable
+    ) {
+        $this->getStockIdForCurrentWebsite = $getStockIdForCurrentWebsite;
+        $this->areProductsSalable = $areProductsSalable;
+    }
+
+    /**
+     * @param Product $product
+     *
+     * @return bool
+     */
+    public function execute(Product $product): bool
+    {
+        $salabilityStatus = false;
+        if ($product->getSku() !== null) {
+            $stockId = $this->getStockIdForCurrentWebsite->execute();
+            if (isset($this->productStatusCache[$stockId][$product->getSku()])) {
+                return $this->productStatusCache[$stockId][$product->getSku()];
+            }
+
+            $stockId = $this->getStockIdForCurrentWebsite->execute();
+            $result = current($this->areProductsSalable->execute([$product->getSku()], $stockId));
+            $salabilityStatus = $result->isSalable();
+            $this->productStatusCache[$stockId][$product->getSku()] = $salabilityStatus;
+        }
+
+        return $salabilityStatus;
+    }
+}
