--- InventoryCatalog/Plugin/CatalogInventory/Helper/Stock/AdaptAssignStatusToProductPlugin.php
@@ -1,110 +1,0 @@
-<?php
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
-declare(strict_types=1);
-
-namespace Magento\InventoryCatalog\Plugin\CatalogInventory\Helper\Stock;
-
-use Magento\Catalog\Model\Product;
-use Magento\CatalogInventory\Helper\Stock;
-use Magento\Framework\Exception\NoSuchEntityException;
-use Magento\InventoryCatalog\Model\GetStockIdForCurrentWebsite;
-use Magento\InventoryCatalogApi\Api\DefaultStockProviderInterface;
-use Magento\InventoryCatalogApi\Model\GetProductIdsBySkusInterface;
-use Magento\InventorySalesApi\Api\AreProductsSalableInterface;
-
-/**
- * Adapt assignStatusToProduct for multi stocks.
- */
-class AdaptAssignStatusToProductPlugin
-{
-    /**
-     * @var GetStockIdForCurrentWebsite
-     */
-    private $getStockIdForCurrentWebsite;
-
-    /**
-     * @var AreProductsSalableInterface
-     */
-    private $areProductsSalable;
-
-    /**
-     * @var DefaultStockProviderInterface
-     */
-    private $defaultStockProvider;
-
-    /**
-     * @var GetProductIdsBySkusInterface
-     */
-    private $getProductIdsBySkus;
-
-    /**
-     * @var array
-     */
-    private $productStatus;
-
-    /**
-     * @param GetStockIdForCurrentWebsite $getStockIdForCurrentWebsite
-     * @param AreProductsSalableInterface $areProductsSalable
-     * @param DefaultStockProviderInterface $defaultStockProvider
-     * @param GetProductIdsBySkusInterface $getProductIdsBySkus
-     */
-    public function __construct(
-        GetStockIdForCurrentWebsite $getStockIdForCurrentWebsite,
-        AreProductsSalableInterface $areProductsSalable,
-        DefaultStockProviderInterface $defaultStockProvider,
-        GetProductIdsBySkusInterface $getProductIdsBySkus
-    ) {
-        $this->getStockIdForCurrentWebsite = $getStockIdForCurrentWebsite;
-        $this->areProductsSalable = $areProductsSalable;
-        $this->defaultStockProvider = $defaultStockProvider;
-        $this->getProductIdsBySkus = $getProductIdsBySkus;
-    }
-
-    /**
-     * Assign stock status to product considering multi stock environment.
-     *
-     * @param Stock $subject
-     * @param callable $proceed
-     * @param Product $product
-     * @param int|null $status
-     * @return void
-     *
-     * @SuppressWarnings(PHPMD.UnusedFormalParameter)
-     */
-    public function aroundAssignStatusToProduct(
-        Stock $subject,
-        callable $proceed,
-        Product $product,
-        $status = null
-    ) {
-        if (null === $product->getSku()) {
-            return;
-        }
-
-        $stockId = $this->getStockIdForCurrentWebsite->execute();
-        if (isset($this->productStatus[$stockId][$product->getSku()])) {
-            $proceed($product, $this->productStatus[$stockId][$product->getSku()]);
-            return;
-        }
-
-        try {
-            $productIds = $this->getProductIdsBySkus->execute([$product->getSku()]);
-            $productId = current($productIds);
-
-            if (null === $status) {
-                $stockId = $this->getStockIdForCurrentWebsite->execute();
-                $result = $this->areProductsSalable->execute([$product->getSku()], $stockId);
-                $result = current($result);
-                $status = (int)$result->isSalable();
-                $this->productStatus[$stockId][$productId] = $status;
-            }
-
-            $proceed($product, $status);
-        } catch (NoSuchEntityException $e) {
-            return;
-        }
-    }
-}
